refactored computation + ros-binding QR Processing nodes
